requirements.txt
python-telegram-bot==13.7
flask>=2.0.0
python-dotenv>=0.19.0  # Для локального тестирования (опционально)
python-dotenv
python-telegram-bot==20.7
flask==2.0.1
